Bug 1569859 - [devtools] Automagically move the toolbox between the original tab we debug and its popups. r=nchevobbe
This is also behind "devtools.popups.debug", to be set to true manually.

This special mode break the old fundamental principal where a given toolbox
is bound to one unique given tab. Now one toolbox starts being shared between
many tabs.

When we select the original tab we debug, or any of its popups opened in distinct tabs,
we will move the toolbox between each of these tabs.
We will have one toolbox instance, one toolbox iframe, which will be moved
around each tab's host. This is somewhat similar to host switching within the same tab.
This is all based on the same trick where we swap the toolbox iframe to another location.

Differential Revision: https://phabricator.services.mozilla.com/D131802

+++ devtools/client/framework/toolbox-host-manager.js
@@ -59,6 +59,16 @@ let ID_COUNTER = 1;
 function ToolboxHostManager(descriptor, hostType, hostOptions) {
   this.descriptor = descriptor;
 
+  // When debugging a local tab, we keep a reference of the current tab into which the toolbox is displayed.
+  // This will only change from the descriptor's localTab when we start debugging popups (i.e. window.open).
+  this.currentTab = this.descriptor.localTab;
+
+  // Keep the previously instantiated Host for all tabs where we displayed the Toolbox.
+  // This will only be useful when we start debugging popups (i.e. window.open).
+  // This is used to re-use the previous host instance when we re-select the original tab
+  // we were debugging before the popup opened.
+  this.hostPerTab = new Map();
+
   this.frameId = ID_COUNTER++;
 
   if (!hostType) {
@@ -69,6 +79,7 @@ function ToolboxHostManager(descriptor, hostType, hostOptions) {
       hostType = Services.prefs.getCharPref(LAST_HOST);
     }
   }
+  this.eventController = new AbortController();
   this.host = this.createHost(hostType, hostOptions);
   this.hostType = hostType;
   this.telemetry = new Telemetry();
@@ -81,11 +92,15 @@ function ToolboxHostManager(descriptor, hostType, hostOptions) {
 ToolboxHostManager.prototype = {
   async create(toolId) {
     await this.host.create();
+    if (this.currentTab) {
+      this.hostPerTab.set(this.currentTab, this.host);
+    }
 
     this.host.frame.setAttribute("aria-label", L10N.getStr("toolbox.label"));
     this.host.frame.ownerDocument.defaultView.addEventListener(
       "message",
-      this._onMessage
+      this._onMessage,
+      { signal: this.eventController.signal }
     );
 
     const msSinceProcessStart = parseInt(
@@ -162,6 +177,9 @@ ToolboxHostManager.prototype = {
       case "switch-host":
         this.switchHost(msg.hostType);
         break;
+      case "switch-host-to-tab":
+        this.switchHostToTab(msg.tabBrowsingContextID);
+        break;
       case "raise-host":
         this.host.raise();
         break;
@@ -178,7 +196,15 @@ ToolboxHostManager.prototype = {
 
   destroy() {
     Services.prefs.removeObserver(ZOOM_VALUE_PREF, this.setMinWidthWithZoom);
+    this.eventController.abort();
+    this.eventController = null;
     this.destroyHost();
+    // When we are debugging popup, we created host for each popup opened
+    // in some other tabs. Ensure destroying them here.
+    for (const host of this.hostPerTab.values()) {
+      host.destroy();
+    }
+    this.hostPerTab.clear();
     this.host = null;
     this.hostType = null;
     this.descriptor = null;
@@ -201,11 +227,27 @@ ToolboxHostManager.prototype = {
     if (!Hosts[hostType]) {
       throw new Error("Unknown hostType: " + hostType);
     }
-    const newHost = new Hosts[hostType](this.descriptor.localTab, options);
+    const newHost = new Hosts[hostType](this.currentTab, options);
     return newHost;
   },
 
-  async switchHost(hostType) {
+  /**
+   * Migrate the toolbox to a new host, while keeping it fully functional.
+   * The toolbox's iframe will be moved as-is to the new host.
+   *
+   * @param {String} hostType
+   *        The new type of host to spawn
+   * @param {Boolean} destroyPreviousHost
+   *        Defaults to true. If false is passed, we will avoid destroying
+   *        the previous host. This is helpful for popup debugging,
+   *        where we migrate the toolbox between two tabs. In this scenario
+   *        we are reusing previously instantiated hosts. This is especially
+   *        useful when we close the current tab and have to have an
+   *        already instantiated host to migrate to. If we don't have one,
+   *        the toolbox iframe will already be destroyed before we have a chance
+   *        to migrate it.
+   */
+  async switchHost(hostType, destroyPreviousHost = true) {
     if (hostType == "previous") {
       // Switch to the last used host for the toolbox UI.
       // This is determined by the devtools.toolbox.previousHost pref.
@@ -234,8 +276,9 @@ ToolboxHostManager.prototype = {
 
     // change toolbox document's parent to the new host
     newIframe.swapFrameLoaders(iframe);
-
-    this.destroyHost();
+    if (destroyPreviousHost) {
+      this.destroyHost();
+    }
 
     if (
       this.hostType !== Toolbox.HostType.BROWSERTOOLBOX &&
@@ -245,11 +288,15 @@ ToolboxHostManager.prototype = {
     }
 
     this.host = newHost;
+    if (this.currentTab) {
+      this.hostPerTab.set(this.currentTab, newHost);
+    }
     this.hostType = hostType;
     this.host.setTitle(this.host.frame.contentWindow.document.title);
     this.host.frame.ownerDocument.defaultView.addEventListener(
       "message",
-      this._onMessage
+      this._onMessage,
+      { signal: this.eventController.signal }
     );
 
     this.setMinWidthWithZoom();
@@ -269,20 +316,51 @@ ToolboxHostManager.prototype = {
   },
 
   /**
+   * When we are debugging popup, we are moving around the toolbox between original tab
+   * and popup tabs. This method will only move the host to a new tab, while
+   * keeping the same host type.
+   *
+   * @param {String} tabBrowsingContextID
+   *        The ID of the browsing context of the tab we want to move to.
+   */
+  async switchHostToTab(tabBrowsingContextID) {
+    const { gBrowser } = this.host.frame.ownerDocument.defaultView;
+
+    const previousTab = this.currentTab;
+    const newTab = gBrowser.tabs.find(
+      tab => tab.linkedBrowser.browsingContext.id == tabBrowsingContextID
+    );
+    // Note that newTab will be undefined when the popup opens in a new top level window.
+    if (newTab && newTab != previousTab) {
+      this.currentTab = newTab;
+      const newHost = this.hostPerTab.get(this.currentTab);
+      if (newHost) {
+        newHost.frame.swapFrameLoaders(this.host.frame);
+        this.host = newHost;
+      } else {
+        await this.switchHost(this.hostType, false);
+      }
+      previousTab.addEventListener(
+        "TabSelect",
+        event => {
+          this.switchHostToTab(event.target.linkedBrowser.browsingContext.id);
+        },
+        { once: true, signal: this.eventController.signal }
+      );
+    }
+
+    this.postMessage({
+      name: "switched-host-to-tab",
+      browsingContextID: tabBrowsingContextID,
+    });
+  },
+
+  /**
    * Destroy the current host, and remove event listeners from its frame.
    *
    * @return {promise} to be resolved when the host is destroyed.
    */
   destroyHost() {
-    // When Firefox toplevel is closed, the frame may already be detached and
-    // the top level document gone
-    if (this.host.frame.ownerDocument.defaultView) {
-      this.host.frame.ownerDocument.defaultView.removeEventListener(
-        "message",
-        this._onMessage
-      );
-    }
-
     return this.host.destroy();
   },
 };
